Refactor CMake files, fix includes, add new method of resolving dependencies and update them
