+++ networking/CMakeLists.txt
@@ -1,5 +1,4 @@
-#include(ccache)
-#include(cotire)
+cmake_minimum_required(VERSION 3.16)
 
 # find the protobuf compiler and libraries
 add_library(networking STATIC
@@ -10,12 +9,12 @@ add_library(networking STATIC
 
 target_include_directories(networking
         PUBLIC include
+        PRIVATE include/networking
         )
 
 target_link_libraries(networking
         PUBLIC stx
-        PUBLIC roboteam_proto
+        PRIVATE roboteam_proto
         # since Publishers/subscribers are templates these 2 libraries need to be PUBLIC
-        PUBLIC lib::zmq
-        PUBLIC lib::zmqpp
+        PUBLIC zmqpp
         )
