+++ networking/include/networking/Publisher.h
